Update project to Meteor 1.3
