+++ .meteor/release
@@ -1,1 +1,1 @@
-0.6.3.1
+METEOR@1.3.4.1
