+++ .meteor/versions
@@ -1,0 +1,67 @@
+accounts-base@1.2.8
+accounts-ui@1.1.9
+accounts-ui-unstyled@1.1.12
+allow-deny@1.0.5
+autopublish@1.0.7
+babel-compiler@6.8.3
+babel-runtime@0.1.9_1
+base64@1.0.9
+binary-heap@1.0.9
+blaze@2.1.8
+blaze-html-templates@1.0.4
+blaze-tools@1.0.9
+boilerplate-generator@1.0.9
+caching-compiler@1.0.5_1
+caching-html-compiler@1.0.6
+callback-hook@1.0.9
+check@1.2.3
+ddp@1.2.5
+ddp-client@1.2.8_1
+ddp-common@1.2.6
+ddp-rate-limiter@1.0.5
+ddp-server@1.2.8_1
+deps@1.0.12
+diff-sequence@1.0.6
+ecmascript@0.4.6_1
+ecmascript-runtime@0.2.11_1
+ejson@1.0.12
+geojson-utils@1.0.9
+html-tools@1.0.10
+htmljs@1.0.10
+id-map@1.0.8
+insecure@1.0.7
+jquery@1.11.9
+less@2.6.3_1
+localstorage@1.0.11
+logging@1.0.13_1
+meteor@1.1.15_1
+minifier-css@1.1.12_1
+minifier-js@1.1.12_1
+minimongo@1.0.17
+modules@0.6.4
+modules-runtime@0.6.4_1
+mongo@1.1.9_1
+mongo-id@1.0.5
+npm-mongo@1.4.44_1
+observe-sequence@1.0.12
+ordered-dict@1.0.8
+promise@0.7.2_1
+random@1.0.10
+rate-limit@1.0.5
+reactive-dict@1.1.8
+reactive-var@1.0.10
+retry@1.0.8
+routepolicy@1.0.11
+service-configuration@1.0.10
+session@1.1.6
+spacebars@1.0.12
+spacebars-compiler@1.0.12
+standard-minifier-css@1.0.7_1
+standard-minifier-js@1.0.7_1
+templating@1.1.12_1
+templating-tools@1.0.4
+tracker@1.0.14
+ui@1.0.11
+underscore@1.0.9
+webapp@1.2.9_1
+webapp-hashing@1.0.9
