+++ admin.html
@@ -9,6 +9,6 @@
 <template name="hello">
   <h1>Hello World!</h1>
   {{greeting}}
-  {{loginButtons}}
+  {{> loginButtons}}
   <input type="button" value="Click" />
 </template>
