+++ admin.js
@@ -1,19 +1,22 @@
-if (Meteor.isClient) {
-  Template.hello.greeting = function () {
-    return "Welcome to admin.";
-  };
+import { Meteor } from 'meteor/meteor';
+import { Template } from 'meteor/templating';
 
-  Template.hello.events({
-    'click input' : function () {
-      // template data, if any, is available in 'this'
-      if (typeof console !== 'undefined')
-        console.log("You pressed the button");
-    }
-  });
+if (Meteor.isClient) {
+    Template.hello.greeting = function() {
+        return "Welcome to admin.";
+    };
+
+    Template.hello.events({
+        'click input': function() {
+            // template data, if any, is available in 'this'
+            if (typeof console !== 'undefined')
+                console.log("You pressed the button");
+        }
+    });
 }
 
 if (Meteor.isServer) {
-  Meteor.startup(function () {
-    // code to run on server at startup
-  });
+    Meteor.startup(function() {
+        // code to run on server at startup
+    });
 }
