--- client/admin.js
@@ -1,9 +1,0 @@
-Meteor.loginAsAdmin = function(password, callback) {
-  
-  var loginRequest = {admin: true, password: password};
-
-  Accounts.callLoginMethod({
-    methodArguments: [loginRequest],
-    userCallback: callback
-  });
-};
